# internal/rest/openapi.yaml
# contracts/rest-openapi.yaml — REST contract for go-rag server APIs
# Feature 003-rest-grpc-api. Hand-authored OpenAPI 3.1 for the internal/rest
# adapter (stdlib net/http). Served by the server at GET /openapi.yaml.
# Mirrors proto/gorag.proto 1:1 (parity). Auth: existing local bearer token
# (Authorization: Bearer <mcp.token>); empty token ⇒ no auth (dev mode).

openapi: 3.1.0
info:
  title: go-rag REST API
  version: 0.1.0
  description: |
    REST projection of go-rag's unified engine. Identical results to the gRPC and
    MCP transports — all three call the same internal/engine facade.
servers:
  - url: http://127.0.0.1:7879

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    QueryRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        k: { type: integer, default: 5, minimum: 1, maximum: 100 }
        mode: { type: string, enum: [hybrid, semantic, keyword], default: hybrid }
        no_rerank: { type: boolean, default: false }
        threshold: { type: number, default: 0.0 }
    QueryHit:
      type: object
      properties:
        chunk_id: { type: string }
        document_id: { type: string }
        score: { type: number }
        content: { type: string }
        file_path: { type: string }
        page: { type: integer }
    QueryResponse:
      type: object
      properties:
        hits: { type: array, items: { $ref: '#/components/schemas/QueryHit' } }
    StatusResponse:
      type: object
      properties:
        documents: { type: integer }
        chunks: { type: integer }
        embeddings: { type: integer }
        dimensions: { type: integer }
        embedding_model: { type: string }
        reranker: { type: string }
        ollama_url: { type: string }
        embeddings_complete: { type: boolean }
    IngestSummary:
      type: object
      properties:
        new: { type: integer }
        skipped: { type: integer }
        modified: { type: integer }
        deleted: { type: integer }
        errors: { type: integer }
    AddRequest:
      type: object
      required: [path]
      properties:
        path: { type: string }
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Liveness/readiness
      security: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  storage_open: { type: boolean }
                  embedder_reachable: { type: boolean }

  /openapi.yaml:
    get:
      summary: This OpenAPI document
      security: []
      responses:
        '200': { description: openapi spec }

  /metrics:
    get:
      summary: Scraped Prometheus metrics (gorag_* latency/count/error families; unauth, loopback)
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain:
              schema:
                type: string
  /v1/query:
    post:
      summary: Hybrid/semantic/keyword search
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryRequest' }
      responses:
        '200':
          description: ranked results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryResponse' }
        '400': { description: bad request }

  /v1/status:
    get:
      summary: Corpus counts + model + reachability
      responses:
        '200':
          description: status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusResponse' }

  /v1/add:
    post:
      summary: Ingest a file or directory (ACK fast, index async)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddRequest' }
      responses:
        '200':
          description: ingest summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestSummary' }

  /v1/scan:
    post:
      summary: Scan watched dirs once; apply changes
      responses:
        '200':
          description: scan summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestSummary' }

  /v1/reprocess:
    post:
      summary: Force re-ingest a path
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddRequest' }
      responses:
        '200':
          description: reprocess summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestSummary' }

  /v1/migrate:
    post:
      summary: Re-embed docs to the current model
      responses:
        '200':
          description: migrate summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestSummary' }

  /v1/migrate/plan:
    post:
      summary: Preview a migration (read-only) — what would change + cost, no re-embedding (H24)
      responses:
        '200':
          description: migration plan
          content:
            application/json:
              schema:
                type: object
                properties:
                  target_model: { type: string }
                  total: { type: integer }
                  stale_total: { type: integer }
                  consistent: { type: boolean }
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        model: { type: string }
                        count: { type: integer }
                        stale: { type: boolean }
                  dimensions:
                    type: array
                    items:
                      type: object
                      properties:
                        dim: { type: integer }
                        count: { type: integer }
                  estimate:
                    type: object
                    properties:
                      stale_embeddings: { type: integer }
                      model_change: { type: boolean }
                      mixed_corpus: { type: boolean }
                      note: { type: string }

  /v1/files:
    get:
      summary: List ingested files
      responses:
        '200':
          description: file list
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        file_path: { type: string }
                        file_type: { type: string }
                        status: { type: string }
                        chunk_count: { type: integer }

  /v1/dirs:
    get:
      summary: List ingested dirs with counts
      responses:
        '200':
          description: dir list
          content:
            application/json:
              schema:
                type: object
                properties:
                  dirs:
                    type: array
                    items:
                      type: object
                      properties:
                        dir: { type: string }
                        files: { type: integer }
                        chunks: { type: integer }

  /v1/documents:
    get:
      summary: List documents — ingested_at cursor + status + tag filter + pagination (spec 039 / 047 R3)
      parameters:
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: page_token
          in: query
          schema: { type: string }
          description: opaque pagination cursor from a prior next_page_token
        - name: after
          in: query
          schema: { type: string, format: date-time }
          description: RFC3339; only docs with ingested_at > after
        - name: status
          in: query
          schema: { type: string, enum: [embedded, pending, error] }
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
          description: spec 047 R3 repeatable match-any tag filter (?tag=a&tag=b → union)
      responses:
        '200':
          description: a page of documents + next_page_token (empty result = empty array)
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents: { type: array, description: "ordered (ingested_at ASC, id ASC); same shape as GetChunk's document" }
                  next_page_token: { type: string, description: "opaque; empty when last page" }
        '400':
          description: invalid-argument (page_size out of range, non-RFC3339 after, unknown status, malformed page_token)

  /v1/documents/{document_id}/chunks:
    get:
      summary: List one document's chunks, paginated (spec 047 / T006)
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: page_token
          in: query
          schema: { type: string }
          description: opaque pagination cursor from a prior next_page_token
      responses:
        '200':
          description: a page of chunks + next_page_token (unknown doc / empty page = empty array, not 404)
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks: { type: array, description: "ordered (chunk_index ASC, chunk_id ASC); same shape as GetChunk's chunk" }
                  next_page_token: { type: string, description: "opaque; empty when last page" }
        '400':
          description: invalid-argument (empty document_id, page_size out of range, malformed page_token)

  /v1/config:
    get:
      summary: Get config (all, or one via ?key=)
      parameters:
        - in: query
          name: key
          schema: { type: string }
      responses:
        '200':
          description: config values
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
    put:
      summary: Set a config value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key: { type: string }
                value: { type: string }
      responses:
        '200':
          description: confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string }
                  value: { type: string }

  /v1/vaults:
    get:
      summary: List all vaults with doc counts
      responses:
        '200':
          description: vault list
          content:
            application/json:
              schema:
                type: object
                properties:
                  vaults:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        documents: { type: integer }
  /v1/poison:
    get:
      summary: List chunks flagged as injection-poisoning (excluded from default results)
      responses:
        '200':
          description: flagged chunks with per-signal verdict breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  flagged:
                    type: array
                    items:
                      type: object
                      properties:
                        chunk_id: { type: string }
                        document_id: { type: string }
                        preview: { type: string }
                        verdict:
                          type: object
                          description: level/score/signals/matched_phrases
  /v1/chunks/{id}:
    get:
      summary: Fetch a single chunk by its content-addressed ID, with parent document metadata (spec 035)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: content-addressed chunk_id
      responses:
        '200':
          description: the chunk and its parent document metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunk: { type: object, description: "chunk_id, document_id, content, page_number, section_context, poisoning, near_dup, …" }
                  document: { type: object, description: "id, content_hash, source_path, file_path, file_type, status, summary, enrichment_status, …" }
        '404':
          description: chunk_id not found in this vault (or belongs to another vault)
        '400':
          description: empty or whitespace-only chunk_id
  /v1/chunks/{id}/context:
    get:
      summary: Fetch a chunk plus up to N neighbours on each side, with the target index and parent document (spec 037)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: content-addressed chunk_id
        - name: window
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 10, default: 2 }
          description: neighbours per side (default 2; 0 returns exactly the target ≡ GetChunk)
      responses:
        '200':
          description: the ordered context window, the target index, and the parent document
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks: { type: array, description: "ordered [predecessors][target][successors]; same shape as GET /v1/chunks/{id}'s chunk" }
                  target_index: { type: integer }
                  document: { type: object, description: "parent document metadata; absent when the chunk is an orphan" }
        '404':
          description: chunk_id not found in this vault (or belongs to another vault)
        '400':
          description: empty or whitespace-only chunk_id, or window out of [0,10]
  /v1/chunks/batch:
    post:
      summary: Resolve up to 100 chunks by id in one call (spec 038)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chunk_ids: { type: array, items: { type: string }, maxItems: 100 }
              required: [chunk_ids]
      responses:
        '200':
          description: one result per requested id, in request order (per-id errors in-band)
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        chunk_id: { type: string }
                        chunk: { type: object, description: "same shape as GET /v1/chunks/{id}'s chunk; absent when not found" }
                        error: { type: string, description: "non-empty iff this chunk_id failed ('not found')" }
                        document: { type: object, description: "parent document metadata; absent when orphan/not-found" }
        '400':
          description: structural invalid-argument — empty list, >100 ids, or an empty/whitespace element (NEVER 404 for a missing id)
  /v1/poison/{id}/release:
    post:
      summary: Release a flagged chunk (false-positive override) — retrievable by default
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: released
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunk_id: { type: string }
                  level: { type: string }
  /v1/poison/{id}/reset:
    post:
      summary: Undo a release — re-quarantines the chunk if its score is flagged
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunk_id: { type: string }
                  status: { type: string }
  /v1/poison/rescan:
    post:
      summary: Re-score the whole corpus against the current detector (idempotent; no re-ingest)
      responses:
        '200':
          description: rescan counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  rescored: { type: integer }
                  flagged: { type: integer }