# internal/rest/server.go
// Package rest is the REST (HTTP/JSON) transport adapter for go-rag, built on
// the Go stdlib net/http ServeMux (zero added dependencies). Like the gRPC and
// MCP adapters, it is a thin projection of the shared internal/engine facade —
// adapters add no independent logic, so REST returns identical results to gRPC
// and MCP.
package rest

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	"github.com/madeinoz67/go-rag/internal/audit"
	"github.com/madeinoz67/go-rag/internal/auth"
	"github.com/madeinoz67/go-rag/internal/engine"
	"github.com/madeinoz67/go-rag/internal/observe"
)

// Server is the REST transport adapter over the engine facade.
type Server struct {
	eng   *engine.Engine
	token string      // empty => auth disabled (local development)
	store *auth.Store // spec 045: credential store for /api/auth/* + unified Validate
}

// New returns a REST adapter backed by eng. token enables bearer auth (matching
// MCP/gRPC); empty disables it for trusted loopback use. The auth store is
// built from eng.DB() when present so /api/auth/* (spec 045) can mint/validate
// sessions and verify the admin password.
func New(eng *engine.Engine, token string) *Server {
	s := &Server{eng: eng, token: token}
	if eng != nil {
		if db := eng.DB(); db != nil {
			s.store = auth.NewStore(db)
		}
	}
	return s
}

// route is one REST route. routes is the authoritative route table: Handler
// wires each entry to its handler, and the OpenAPI parity test (T035) asserts
// this table matches the served spec exactly. Add a route here AND to
// openapi.yaml; the test catches drift in either direction.
type route struct {
	method string
	path   string
	auth   bool // true ⇒ bearer-guarded
}

var routes = []route{
	{"GET", "/health", false},
	{"GET", "/metrics", false}, // H17/spec 020: scraped Prometheus endpoint (unauth, loopback)
	{"GET", "/openapi.yaml", false},
	{"POST", "/v1/query", true},
	{"GET", "/v1/status", true},
	{"POST", "/v1/add", true},
	{"POST", "/v1/scan", true},
	{"POST", "/v1/reprocess", true},
	{"POST", "/v1/migrate", true},
	{"POST", "/v1/migrate/plan", true}, // H24/spec 028: read-only migration preview
	{"GET", "/v1/files", true},
	{"GET", "/v1/dirs", true},
	{"GET", "/v1/documents", true},                      // spec 039: list documents (cursor + status + tags + pagination)
	{"GET", "/v1/documents/{document_id}/chunks", true}, // spec 047 T006: list a document's chunks (cursor pagination)
	{"GET", "/v1/config", true},
	{"PUT", "/v1/config", true},
	{"GET", "/v1/vaults", true},
	{"GET", "/v1/poison", true},               // H04/spec 019: list flagged chunks
	{"POST", "/v1/poison/{id}/release", true}, // H04/spec 019: false-positive override
	{"POST", "/v1/poison/{id}/reset", true},   // H04/spec 019: undo a release
	{"POST", "/v1/poison/rescan", true},       // H04/spec 019: re-score the corpus
	{"GET", "/v1/chunks/{id}", true},          // spec 035: fetch a chunk by content-addressed ID
	{"GET", "/v1/chunks/{id}/context", true},  // spec 037: a chunk plus up to N neighbours each side
	{"POST", "/v1/chunks/batch", true},        // spec 038: resolve up to 100 chunks by id
}

// Handler returns the http.Handler serving the REST API (Go 1.22+ pattern mux),
// built from the routes table.
func (s *Server) Handler() http.Handler {
	mux := http.NewServeMux()
	for _, r := range routes {
		h := s.handlerFor(r.method, r.path)
		if h == nil {
			panic(fmt.Sprintf("rest: no handler wired for %s %s", r.method, r.path))
		}
		if r.auth {
			h = s.guard(h)
		}
		mux.HandleFunc(r.method+" "+r.path, h)
	}
	s.mountAuth(mux) // spec 045: /api/auth/* (login + session management)
	return mux
}

// handlerFor resolves a route table entry to its handler method.
func (s *Server) handlerFor(method, path string) http.HandlerFunc {
	switch method + " " + path {
	case "GET /health":
		return s.handleHealth
	case "GET /metrics":
		return s.handleMetrics
	case "GET /openapi.yaml":
		return s.handleOpenAPI
	case "POST /v1/query":
		return s.handleQuery
	case "GET /v1/status":
		return s.handleStatus
	case "POST /v1/add":
		return s.handleAdd
	case "POST /v1/scan":
		return s.handleScan
	case "POST /v1/reprocess":
		return s.handleReprocess
	case "POST /v1/migrate":
		return s.handleMigrate
	case "POST /v1/migrate/plan":
		return s.handleMigratePlan // H24/spec 028
	case "GET /v1/files":
		return s.handleFiles
	case "GET /v1/dirs":
		return s.handleDirs
	case "GET /v1/documents":
		return s.handleListDocuments // spec 039
	case "GET /v1/documents/{document_id}/chunks":
		return s.handleListChunks // spec 047 T006
	case "GET /v1/config":
		return s.handleConfigGet
	case "PUT /v1/config":
		return s.handleConfigSet
	case "GET /v1/vaults":
		return s.handleVaults
	case "GET /v1/poison":
		return s.handlePoisonList
	case "POST /v1/poison/{id}/release":
		return s.handlePoisonRelease
	case "POST /v1/poison/{id}/reset":
		return s.handlePoisonReset
	case "POST /v1/poison/rescan":
		return s.handlePoisonRescan
	case "GET /v1/chunks/{id}":
		return s.handleGetChunk // spec 035
	case "GET /v1/chunks/{id}/context":
		return s.handleGetChunkContext // spec 037
	case "POST /v1/chunks/batch":
		return s.handleBatchGetChunks // spec 038
	}
	return nil
}

// principalCtxKey is the context key under which guard stores the authenticated
// Principal. Handlers read it via PrincipalFromContext to enforce Mode.
type principalCtxKey struct{}

// PrincipalFromContext returns the authenticated Principal guard placed on the
// request context. ok is false when no Principal is present (e.g. an
// unauthenticated route).
func PrincipalFromContext(ctx context.Context) (auth.Principal, bool) {
	p, ok := ctx.Value(principalCtxKey{}).(auth.Principal)
	return p, ok
}

// guard wraps a handler with unified bearer auth (spec 045 US2). All transports
// delegate to the single auth.Validate so a key works identically on
// REST/gRPC/MCP and a revoke propagates to all three. The loopback bypass
// (US5) is handled inside Validate: local dev with no credentials minted just
// works; the moment a credential exists, every request must present it.
func (s *Server) guard(h http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if s.store == nil { // test-only: engine without a backing DB
			h(w, r)
			return
		}
		p, err := s.store.Validate(r)
		if err != nil {
			audit.Log(audit.AuthFailEvent("rest", "missing or invalid bearer")) // H18 audit
			writeError(w, http.StatusUnauthorized, "unauthorized")
			return
		}
		h(w, r.WithContext(context.WithValue(r.Context(), principalCtxKey{}, p)))
	}
}

// handleHealth is the unified liveness/readiness probe (GET /health). Reports
// liveness (ok) + readiness (ready) — ready is false on hard embedding drift
// (audit H11/spec 017), so clients/orchestrators reading the body do not route
// query traffic; ok stays true while the process is up. Identical to the gRPC
// Health RPC (both call engine.Health). Unauthenticated so probes don't need a
// token. HTTP stays 200 on drift (liveness) — readiness is in the body, not the
// status code, to avoid restart-loops if /health is used as a liveness probe.
func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	h := s.eng.Health(r.Context())
	writeJSON(w, http.StatusOK, map[string]any{
		"ok":                 h.OK,
		"ready":              h.Ready,
		"storage_open":       h.StorageOpen,
		"embedder_reachable": h.EmbedderReachable,
		"drift_verdict":      h.DriftVerdict,
	})
}

// handleMetrics serves the scraped Prometheus /metrics endpoint (H17/spec 020).
// Unauthenticated + loopback-only (like /health); disabled when metrics_enabled=false.
func (s *Server) handleMetrics(w http.ResponseWriter, r *http.Request) {
	if !s.eng.Config().EffectiveMetricsEnabled() {
		http.NotFound(w, r)
		return
	}
	observe.MetricsHandler().ServeHTTP(w, r)
}

// --- helpers ---

func writeJSON(w http.ResponseWriter, code int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	_ = json.NewEncoder(w).Encode(v)
}

func writeError(w http.ResponseWriter, code int, msg string) {
	writeJSON(w, code, map[string]any{"error": msg})
}

// writeEngineErr maps an engine error to the right HTTP status: client-input
// errors (ErrInvalid) → 400, everything else (storage/index/embedder faults) →
// 500. nil is a no-op.
func writeEngineErr(w http.ResponseWriter, err error) {
	if err == nil {
		return
	}
	if errors.Is(err, engine.ErrInvalid) {
		writeError(w, http.StatusBadRequest, err.Error())
		return
	}
	if errors.Is(err, engine.ErrNotFound) { // spec 035: a missing id is a normal client outcome (FR-002)
		writeError(w, http.StatusNotFound, err.Error())
		return
	}
	writeError(w, http.StatusInternalServerError, err.Error())
}

// checkBearer was the bespoke per-transport bearer check deleted in spec 045
// US2 — every transport now delegates to auth.Validate. (Symbol removed.)